Refactor `file_ops` into `core` module for better organization.
Moved `file_ops` functionality into a new `core` module to improve structure and maintainability. Updated all references to reflect the new path and added an `__init__.py` file in the `core` directory for proper module initialization.
