+++ dev_kit_mcp_server/tools/__init__.py
@@ -1,6 +1,6 @@
 """Tools subpackage for MCP server implementations."""
 
-from .file_ops import FileOperation
+from .core import FileOperation
 from .file_sys.create import CreateDirOperation
 from .file_sys.move import MoveDirOperation
 from .file_sys.remove import RemoveFileOperation
