+++ dev_kit_mcp_server/tools/file_sys/create.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Callable, Dict
 
-from ..file_ops import FileOperation
+from ..core.file_ops import FileOperation
 
 
 @dataclass
